(NFC) Add intrinsics.cpp
Introduce a separate source file that contains code processing LLVM
intrinsics usage in ISPC source code (--enable-llvm-intrinsics).
