+++ src/builtins.cpp
@@ -56,138 +56,6 @@
 using namespace ispc;
 using namespace ispc::builtin;
 
-/** Given an LLVM type, try to find the equivalent ispc type.  Note that
-    this is an under-constrained problem due to LLVM's type representations
-    carrying less information than ispc's.  (For example, LLVM doesn't
-    distinguish between signed and unsigned integers in its types.)
-
-    Because this function is only used for generating ispc declarations of
-    functions defined in LLVM bitcode in the builtins-*.ll files, in practice
-    we can get enough of what we need for the relevant cases to make things
-    work, partially with the help of the intAsUnsigned parameter, which
-    indicates whether LLVM integer types should be treated as being signed
-    or unsigned.
-
- */
-static const Type *lLLVMTypeToISPCType(const llvm::Type *t, bool intAsUnsigned) {
-    if (t == LLVMTypes::VoidType) {
-        return AtomicType::Void;
-    }
-
-    // uniform
-    else if (t == LLVMTypes::BoolType) {
-        return AtomicType::UniformBool;
-    } else if (t == LLVMTypes::Int8Type) {
-        return intAsUnsigned ? AtomicType::UniformUInt8 : AtomicType::UniformInt8;
-    } else if (t == LLVMTypes::Int16Type) {
-        return intAsUnsigned ? AtomicType::UniformUInt16 : AtomicType::UniformInt16;
-    } else if (t == LLVMTypes::Int32Type) {
-        return intAsUnsigned ? AtomicType::UniformUInt32 : AtomicType::UniformInt32;
-    } else if (t == LLVMTypes::Float16Type) {
-        return AtomicType::UniformFloat16;
-    } else if (t == LLVMTypes::FloatType) {
-        return AtomicType::UniformFloat;
-    } else if (t == LLVMTypes::DoubleType) {
-        return AtomicType::UniformDouble;
-    } else if (t == LLVMTypes::Int64Type) {
-        return intAsUnsigned ? AtomicType::UniformUInt64 : AtomicType::UniformInt64;
-    }
-
-    // varying
-    if (t == LLVMTypes::Int8VectorType) {
-        return intAsUnsigned ? AtomicType::VaryingUInt8 : AtomicType::VaryingInt8;
-    } else if (t == LLVMTypes::Int16VectorType) {
-        return intAsUnsigned ? AtomicType::VaryingUInt16 : AtomicType::VaryingInt16;
-    } else if (t == LLVMTypes::Int32VectorType) {
-        return intAsUnsigned ? AtomicType::VaryingUInt32 : AtomicType::VaryingInt32;
-    } else if (t == LLVMTypes::Float16VectorType) {
-        return AtomicType::VaryingFloat16;
-    } else if (t == LLVMTypes::FloatVectorType) {
-        return AtomicType::VaryingFloat;
-    } else if (t == LLVMTypes::DoubleVectorType) {
-        return AtomicType::VaryingDouble;
-    } else if (t == LLVMTypes::Int64VectorType) {
-        return intAsUnsigned ? AtomicType::VaryingUInt64 : AtomicType::VaryingInt64;
-    } else if (t == LLVMTypes::MaskType) {
-        return AtomicType::VaryingBool;
-    }
-
-    // pointers to uniform
-    else if (t == LLVMTypes::Int8PointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::UniformUInt8 : AtomicType::UniformInt8);
-    } else if (t == LLVMTypes::Int16PointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::UniformUInt16 : AtomicType::UniformInt16);
-    } else if (t == LLVMTypes::Int32PointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::UniformUInt32 : AtomicType::UniformInt32);
-    } else if (t == LLVMTypes::Int64PointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::UniformUInt64 : AtomicType::UniformInt64);
-    } else if (t == LLVMTypes::Float16PointerType) {
-        return PointerType::GetUniform(AtomicType::UniformFloat16);
-    } else if (t == LLVMTypes::FloatPointerType) {
-        return PointerType::GetUniform(AtomicType::UniformFloat);
-    } else if (t == LLVMTypes::DoublePointerType) {
-        return PointerType::GetUniform(AtomicType::UniformDouble);
-    }
-
-    // pointers to varying
-    else if (t == LLVMTypes::Int8VectorPointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::VaryingUInt8 : AtomicType::VaryingInt8);
-    } else if (t == LLVMTypes::Int16VectorPointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::VaryingUInt16 : AtomicType::VaryingInt16);
-    } else if (t == LLVMTypes::Int32VectorPointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::VaryingUInt32 : AtomicType::VaryingInt32);
-    } else if (t == LLVMTypes::Int64VectorPointerType) {
-        return PointerType::GetUniform(intAsUnsigned ? AtomicType::VaryingUInt64 : AtomicType::VaryingInt64);
-    } else if (t == LLVMTypes::Float16VectorPointerType) {
-        return PointerType::GetUniform(AtomicType::VaryingFloat16);
-    } else if (t == LLVMTypes::FloatVectorPointerType) {
-        return PointerType::GetUniform(AtomicType::VaryingFloat);
-    } else if (t == LLVMTypes::DoubleVectorPointerType) {
-        return PointerType::GetUniform(AtomicType::VaryingDouble);
-    }
-
-    return nullptr;
-}
-
-Symbol *ispc::CreateISPCSymbolForLLVMIntrinsic(llvm::Function *func, SymbolTable *symbolTable) {
-    Symbol *existingSym = symbolTable->LookupIntrinsics(func);
-    if (existingSym != nullptr) {
-        return existingSym;
-    }
-    SourcePos noPos;
-    noPos.name = "LLVM Intrinsic";
-    const llvm::FunctionType *ftype = func->getFunctionType();
-    std::string name = std::string(func->getName());
-    const Type *returnType = lLLVMTypeToISPCType(ftype->getReturnType(), false);
-    if (returnType == nullptr) {
-        Error(SourcePos(),
-              "Return type not representable for "
-              "Intrinsic %s.",
-              name.c_str());
-        // return type not representable in ispc -> not callable from ispc
-        return nullptr;
-    }
-    llvm::SmallVector<const Type *, 8> argTypes;
-    for (unsigned int j = 0; j < ftype->getNumParams(); ++j) {
-        const llvm::Type *llvmArgType = ftype->getParamType(j);
-        const Type *type = lLLVMTypeToISPCType(llvmArgType, false);
-        if (type == nullptr) {
-            Error(SourcePos(),
-                  "Type of parameter %d not "
-                  "representable for Intrinsic %s",
-                  j, name.c_str());
-            return nullptr;
-        }
-        argTypes.push_back(type);
-    }
-    FunctionType *funcType = new FunctionType(returnType, argTypes, noPos);
-    Debug(noPos, "Created Intrinsic symbol \"%s\" [%s]\n", name.c_str(), funcType->GetString().c_str());
-    Symbol *sym = new Symbol(name, noPos, Symbol::SymbolKind::Function, funcType);
-    sym->function = func;
-    symbolTable->AddIntrinsics(sym);
-    return sym;
-}
-
 /** In many of the builtins-*.ll files, we have declarations of various LLVM
     intrinsics that are then used in the implementation of various target-
     specific functions.  This function loops over all of the intrinsic
