+++ src/module.cpp
@@ -457,85 +457,6 @@ int Module::CompileFile() {
     return errorCount;
 }
 
-Symbol *Module::AddLLVMIntrinsicDecl(const std::string &name, ExprList *args, SourcePos pos) {
-    if (g->enableLLVMIntrinsics == false) {
-        Error(SourcePos(), "Calling LLVM intrinsics from ISPC source code is an experimental feature,"
-                           " which can be enabled by passing \"--enable-llvm-intrinsics\" switch to the compiler.\n");
-        return nullptr;
-    }
-
-    llvm::Function *funcDecl = nullptr;
-#ifdef ISPC_XE_ENABLED
-    if (g->target->isXeTarget()) {
-        llvm::GenXIntrinsic::ID ID = llvm::GenXIntrinsic::lookupGenXIntrinsicID(name);
-        if (ID == llvm::GenXIntrinsic::not_any_intrinsic) {
-            Error(pos, "LLVM intrinsic \"%s\" not supported.", name.c_str());
-            return nullptr;
-        }
-        std::vector<llvm::Type *> exprType;
-        Assert(args);
-        int nInits = args->exprs.size();
-        if (llvm::GenXIntrinsic::isOverloadedRet(ID) || llvm::GenXIntrinsic::isOverloadedArg(ID, nInits)) {
-            for (int i = 0; i < nInits; ++i) {
-                const Type *argType = (args->exprs[i])->GetType();
-                Assert(argType);
-                exprType.push_back(argType->LLVMType(g->ctx));
-            }
-        }
-        llvm::ArrayRef<llvm::Type *> argArr(exprType);
-        funcDecl = llvm::GenXIntrinsic::getGenXDeclaration(module, ID, argArr);
-        if (funcDecl) {
-            // ReadNone, ReadOnly and WriteOnly are not supported for intrinsics anymore:
-            FixFunctionAttribute(*funcDecl, llvm::Attribute::ReadNone, llvm::MemoryEffects::none());
-            FixFunctionAttribute(*funcDecl, llvm::Attribute::ReadOnly, llvm::MemoryEffects::readOnly());
-            FixFunctionAttribute(*funcDecl, llvm::Attribute::WriteOnly, llvm::MemoryEffects::writeOnly());
-        }
-    }
-#endif
-    if (!g->target->isXeTarget()) {
-        llvm::TargetMachine *targetMachine = g->target->GetTargetMachine();
-        const llvm::TargetIntrinsicInfo *TII = targetMachine->getIntrinsicInfo();
-#if ISPC_LLVM_VERSION >= ISPC_LLVM_20_0
-        llvm::Intrinsic::ID ID = llvm::Intrinsic::lookupIntrinsicID(llvm::StringRef(name));
-#else
-        llvm::Intrinsic::ID ID = llvm::Function::lookupIntrinsicID(llvm::StringRef(name));
-#endif
-        if (ID == llvm::Intrinsic::not_intrinsic && TII) {
-            ID = static_cast<llvm::Intrinsic::ID>(TII->lookupName(llvm::StringRef(name)));
-        }
-        if (ID == llvm::Intrinsic::not_intrinsic) {
-            Error(pos, "LLVM intrinsic \"%s\" not supported.", name.c_str());
-            return nullptr;
-        }
-        std::vector<llvm::Type *> exprType;
-        if (llvm::Intrinsic::isOverloaded(ID)) {
-            Assert(args);
-            int nInits = args->exprs.size();
-            for (int i = 0; i < nInits; ++i) {
-                const Type *argType = (args->exprs[i])->GetType();
-                Assert(argType);
-                exprType.push_back(argType->LLVMType(g->ctx));
-            }
-        }
-        llvm::ArrayRef<llvm::Type *> argArr(exprType);
-        // Add intrinsic declaration to the module if it's not added yet.
-#if ISPC_LLVM_VERSION >= ISPC_LLVM_20_0
-        funcDecl = llvm::Intrinsic::getOrInsertDeclaration(module, ID, argArr);
-#else
-        funcDecl = llvm::Intrinsic::getDeclaration(module, ID, argArr);
-#endif
-        llvm::StringRef funcName = funcDecl->getName();
-
-        if (g->target->checkIntrinsticSupport(funcName, pos) == false) {
-            return nullptr;
-        }
-    }
-
-    Assert(funcDecl != nullptr);
-    Symbol *funcSym = CreateISPCSymbolForLLVMIntrinsic(funcDecl, symbolTable);
-    return funcSym;
-}
-
 void Module::AddTypeDef(const std::string &name, const Type *type, SourcePos pos) {
     // Typedefs are easy; just add the mapping between the given name and
     // the given type.
